Adding classes for managing data and reports
In this parts the CarbonFootprint class was added with methods to to calculate emissions from energy usage, waste, travel and total. Then the ReportGenerator class was added to generate and save the reports.
The link to the other page if for now deleted and will be added in the next steps.